requirements.txt
python-dotenv==1.0.0
langchain==0.1.1
langchain-openai==0.0.2
openai==1.3.5
pydantic==2.5.2